Add something

+++ CalcStockTax/DbSRV/Models/Investment.cs
@@ -4,10 +4,10 @@
     {
         public int InvestmentId { get; set; }
         public string StockName { get; set; }
-        public int BuyPrice { get; set; }
+        public double BuyPrice { get; set; }
         public int Quantity { get; set; }
         public DateTime BuyDate { get; set; }
-        public int CurrentPrice { get; set; }
+        public double CurrentPrice { get; set; }
 
     }
 }
